+++ src/wai/annotations/core/__init__.py
@@ -10,7 +10,7 @@ from ._InternalFormatConverter import InternalFormatConverter
 from ._JSONWriter import JSONWriter
 from ._Reader import Reader
 from ._SeparateImageWriter import SeparateImageWriter
-from ._Settings import Settings, get_settings, set_settings
+from ._LibrarySettings import LibrarySettings, get_settings, set_settings
 from ._typing import InternalFormat
 from ._Writer import Writer
 from ._ZeroAreaDiscarder import ZeroAreaDiscarder