+++ src/wai/annotations/core/utils/_get_associated_image.py
@@ -11,7 +11,7 @@ def get_associated_image(filename: str) -> Optional[str]:
     :return:            The filename of the found image, None if not found.
     """
     # Check each variation of the extension for each format
-    from .._Settings import get_settings
+    from .._LibrarySettings import get_settings
     for image_format in get_settings().IMAGE_FORMAT_PREFERENCE_ORDER:
         for extension in image_format.value:
             # Get the hypothetical filename for an image of this format